remove function get_element, get_element_from_data_table, wait_element_for_status in action_android and action_web
updated
