+++ features/steps/execure_script.py
@@ -1,6 +1,4 @@
-import configparser
 from behave import *
-from selenium import webdriver
 from Utilities.action_web import ManagementFile
 from ManagementElements.Page import Page
 from ManagementElements.Elements import Elements
@@ -22,11 +20,9 @@ def launchBrowser(context, index):
     array_url = context.url.split(",")
     context.driver.get(array_url[int(index) - 1].strip())
 
-
 @step(u'I open application')
 def step_impl(context):
     print("run application")
-
 
 @step(u'I change the page spec to {page}')
 def change_page(context, page):
@@ -34,41 +30,29 @@ def change_page(context, page):
     page = ManagementFile().read_yaml_file(path_file + "/" + page + ".yaml", dict_page, page)
     context.page_present = page
     return context.page_present
-
-
 @step(u'I click element {element}')
 def click_action(context, element):
-    context.element_page = ManagementFile().get_element(context.page_present, element,
-                                                        context.device.get_platform_name(), context.dict_save_value)
+    context.element_page = common_device().get_element(context.page_present, element,
+                                                       context.device.get_platform_name(), context.dict_save_value)
     common_device().action_page(context.element_page, "click", context.driver, "", context.wait,
-                                    context.dict_save_value, context.device)
-
-
-
+                                context.dict_save_value, context.device)
 @step(u'I type "{text}" into element {element}')
 def type_action(context, text, element):
-    context.element_page = ManagementFile().get_element(context.page_present, element,
-                                                        context.device.get_platform_name(), context.dict_save_value)
+    context.element_page = common_device().get_element(context.page_present, element,
+                                                       context.device.get_platform_name(), context.dict_save_value)
     common_device().action_page(context.element_page, "type", context.driver, text, context.wait,
-                                    context.dict_save_value, context.device)
-
-
+                                context.dict_save_value, context.device)
 @step(u'I wait for element {element} to be {status}')
 def wait_element(context, element, status):
-    context.element_page = ManagementFile().get_element(context.page_present, element,
-                                                        context.device.get_platform_name(), context.dict_save_value)
-    if context.device.get_platform_name() == "WEB":
-        ManagementFile().wait_element_for_status(context.element_page, status, context.driver, context.wait)
-    elif context.device.get_platform_name() == "ANDROID":
-        ManagementFileAndroid().wait_element_for_status(context.element_page, status, context.driver, context.device)
-
-
+    context.element_page = common_device().get_element(context.page_present, element,
+                                                       context.device.get_platform_name(), context.dict_save_value)
+    common_device().wait_element_for_status(context.element_page, status, context.driver, context.device)
 @step(u'I wait for elements with below status')
 def step_impl(context):
-    if context.device.get_platform_name() == "WEB":
-        context.element_page = ManagementFile().get_element_from_data_table(context.page_present, context.table, context.device.get_platform_name(), context.dict_save_value, context.driver, context.wait)
-    elif context.device.get_platform_name() == "ANDROID":
-        ManagementFileAndroid().wait_element_for_status(context.element_page, context.table, context.driver, context.device)
+    context.element_page = common_device().verify_elements_with_status(context.page_present, context.table,
+                                                                       context.device.get_platform_name(),
+                                                                       context.dict_save_value, context.driver,
+                                                                       context.device)
 @step(u'I perform {action} action')
 def step_impl(context, action):
     if context.device.get_platform_name() == "WEB":
@@ -76,8 +60,6 @@ def step_impl(context, action):
     elif context.device.get_platform_name() == "ANDROID":
         ManagementFileAndroid().execute_action_android(context.page_present, action, context.driver,
                                                        context.device.get_wait(), None, None)
-
-
 @step(u'I perform {action} action with override values')
 def step_impl(context, action):
     if context.device.get_platform_name() == "WEB":
@@ -87,20 +69,16 @@ def step_impl(context, action):
         ManagementFileAndroid().execute_action_android(context.page_present, action, context.driver,
                                                        context.device.get_wait(), context.table,
                                                        context.dict_save_value)
-
-
 @step(u'I clear text from element {element}')
 def step_impl(context, element):
-    context.element_page = ManagementFile().get_element(context.page_present, element,
+    context.element_page = common_device().get_element(context.page_present, element,
                                                         context.device.get_platform_name(), context.dict_save_value)
     common_device().action_page(context.element_page, "clear", context.driver, "", context.wait,
-                                    context.dict_save_value, context.device)
-
-
+                                context.dict_save_value, context.device)
 @step(u'I save text for element {element} with key "{key}"')
 def step_impl(context, element, key):
-    context.element_page = ManagementFile().get_element(context.page_present, element,
-                                                        context.device.get_platform_name(), context.dict_save_value)
+    context.element_page = common_device().get_element(context.page_present, element,
+                                                       context.device.get_platform_name(), context.dict_save_value)
     if context.device.get_platform_name() == "WEB":
         context.dict_save_value = ManagementFile().save_text_from_element(context.element_page, context.driver, key,
                                                                           context.dict_save_value, context.wait)
@@ -109,10 +87,7 @@ def step_impl(context, element, key):
                                                                                          context.driver, key,
                                                                                          context.dict_save_value,
                                                                                          context.wait)
-        # print(context.dict_save_value)
     return context.dict_save_value
-
-
 @step(u'I wait 5 seconds')
 def step_impl(context):
     print("wait")
